src/shabd: Add structs Akshar and Shabd
Struct Akshar represents the basic atomic unit of sound that can be
spoken. It essentially is a collection of Varns, and its matra is sum
total of the matras of the varns constituting an Akshar.
Struct Shabd is collection of Akshars that form a meaningful word.

Add Akshars and Shabds and functions to make them from a given
collection of Varns. Add the tests in tests/shabd_tests.rs.

Signed-off-by: Ankit Nautiyal <akn_github@outlook.com>