src/varn: Fix the chihn processing
Currently the analysis for shabd with more than one chihn only uses
the matra of the first chihn. This messes up the analysis with
certain chihns.
Add the matras of consecutive chihns in an akshar. Clamp the sum to 2,
as its the maximum matra the consecutive chihns can contribute to an
akshar.

Signed-off-by: Ankit Nautiyal <akn_github@outlook.com>